Remove Form and replace on Home page with Add Album. Move AddMessage from Roulette to Home page. Modify RouletteProps type. Add font Lexend Deca as main font. Update CSS on various components.
