+++ package.json
@@ -14,17 +14,17 @@
     "@types/react": "18.0.38",
     "@types/react-dom": "18.0.11",
     "eslint": "8.39.0",
-    "eslint-config-next": "13.3.1",
     "firebase": "^9.23.0",
-    "next": "^13.4.19",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.5.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-firebase-hooks": "^5.1.1",
     "reactfire": "^4.2.3",
     "styled-components": "^5.3.10",
     "typescript": "5.0.4"
   },
   "devDependencies": {
-    "@types/styled-components": "^5.1.26"
+    "@types/styled-components": "^5.1.26",
+    "eslint-config-next": "^13.5.3"
   }
 }
