+++ src/app/fonts.ts
@@ -1,0 +1,6 @@
+import { Lexend_Deca } from "next/font/google";
+
+export const lexend_deca = Lexend_Deca({
+  subsets: ["latin"],
+  display: "swap",
+});
